requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
torch
torchvision
pillow
python-multipart
